backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.9
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.1.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.20
httpx==0.27.2
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.0
openpyxl==3.1.2
odfpy==1.4.1
docx2txt==0.8
pytest==7.4.4
pytest-asyncio==0.23.3